generic public key verification and validation
Implements a generic public key structure to be able to allow
different algorithms for public keys other than Edwards 25519.
This also implements the logic to verify the signatures of the hashes,
and then verify the hashes by hashing all contents and ensuring it matches
with one of the validated hashes.