implement keypair and generic way of signing
The keypair is essentially a small wrapping struct over Ed25519's
KeyPair with some addition functions to allow for easy constructing
of the generic `Signer` interface this commit adds also.
It also allows to construct a public key from the keypair wich can
then be used to verify said signatures.