correctly write the 128leb-encoded length
Previously we used the fixed-version of leb128 encoding,
which stores it in all 5 bytes. But since we were slicing the
array shorter, the actual value was corrupt in the binary.
We fixed this using a newly implemented function that still
writes it in the shortest encoding possible, without having to
use a writer.